Update socket references to be int32 to match change in host driver 1.14 socket connect result.

+++ Adafruit_CC3000.h
@@ -123,7 +123,7 @@ class Adafruit_CC3000_Client : public Client {
   int16_t bufsiz;
 
  private:
-  int16_t _socket;
+  int32_t _socket;
 
 };
